requirements.txt
Click==8.1.3
click-repl==0.2.0
grpcio==1.47.0
grpcio-tools==1.47.0
PyYAML==6.0
